+++ HISTORY.rst
@@ -22,8 +22,16 @@ unreleased
 
 **Device Update**
 
-* The Device Update control plane or infrastructure command groups `az iot device-update account` and
+* The Device Update control plane (or infrastructure related) command groups `az iot device-update account` and
   `az iot device-update instance` now use the GA API version of 2022-10-01.
+* The Device Update data plane command groups `az iot device-update device` and
+  `az iot device-update update` now use the GA API version of 2022-10-01.
+* The command `az iot device-update device class list` adds support for `--filter` when no `--group-id` is provided.
+* The parameters `--account`, `--instance`, and `--resource-group` support setting default overridable values via config.
+  Use `az config set` i.e. `az config set defaults.adu_account=<name>` or `az configure` i.e. `az configure --defaults adu_account=<name>`.
+* Introducing the experimental command `az iot device-update update init v5` for initializing (or generating) an import manifest
+  with the desired state.
+* Improved built-in documentation.
 
 
 0.17.0